exercise 4.18-4.22

+++ part4/blog-backend/controllers/blogs.js
@@ -1,8 +1,7 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
-// init
 blogsRouter.get('/', async (request, response) => {
 	const blogs = await Blog.find({}).populate('user', {
 		username: 1,
@@ -21,25 +20,34 @@ blogsRouter.get('/:id', async (request, response) => {
 });
 
 blogsRouter.post('', async (request, response) => {
-	const body = request.body;
-	const user = await User.findById(body.userId);
-	const blog = new Blog({
-		title: body.title,
-		author: body.author,
-		url: body.url,
-		likes: body.likes,
-		user: user.id,
-	});
-
-	const savedBlog = await blog.save();
-	user.blogs = user.blogs.concat(savedBlog._id);
-	await user.save();
-	response.status(201).json(savedBlog);
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (decodedToken.id === request.user.id.toString()) {
+		const body = request.body;
+		const user = request.user;
+		const blog = new Blog({
+			title: body.title,
+			author: body.author,
+			url: body.url,
+			likes: body.likes,
+			user: user.id,
+		});
+		const savedBlog = await blog.save();
+		user.blogs = user.blogs.concat(savedBlog._id);
+		await user.save();
+		response.status(201).json(savedBlog);
+	} else {
+		response.status(401).json({ error: 'token invalid' });
+	}
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-	await Blog.findByIdAndDelete(request.params.id);
-	response.status(204).end();
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (decodedToken.id === request.user.id.toString()) {
+		await Blog.findByIdAndDelete(request.params.id);
+		response.status(204).end();
+	} else {
+		response.status(401).json({ error: 'token invalid' });
+	}
 });
 
 blogsRouter.put('/:id', async (request, response) => {
